# docs/swagger.yaml
definitions:
...
  ds.Ship:
    properties:
      capacity:
        type: number
      containers:
        type: integer
      cranes:
        type: integer
      description:
        type: string
      draft:
        type: number
      length:
        type: number
      name:
        type: string
      photoURL:
        type: string
      shipID:
        type: integer
      width:
        type: number
    type: object
...
info:
  contact: {}
paths:
...
  /api/request_ship/{id}/ships/{ship_id}:
    delete:
      description: Remove a ship from a specific request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ship ID
        in: path
        name: ship_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'description: string'
          schema:
            type: object
        "400":
          description: 'status: string, description: string'
          schema:
            type: object
        "500":
          description: 'status: string, description: string'
          schema:
            type: object
      summary: Удаление корабля из заявки
      tags:
      - request_ships
    put:
      consumes:
      - application/json
      description: Update the number of ships in a specific request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ship ID
        in: path
        name: ship_id
        required: true
        type: integer
      - description: Updated ship count
        in: body
        name: request
        required: true
        schema:
          properties:
            ships_count:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'status: string, message: string'
          schema:
            type: object
        "400":
          description: 'description: string'
          schema:
            type: object
        "500":
          description: 'error: string'
          schema:
            type: object
      summary: Обновление количества кораблей в заявке
      tags:
      - request_ships
...
  /api/ships:
    get:
      description: Retrieve a list of ships with optional filters
      parameters:
      - description: Ship name filter
        in: query
        name: name
        type: string
      - description: Minimum capacity filter
        in: query
        name: capacity
        type: string
      - description: Active status filter
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 'data: []ds.Ship, count: int'
          schema:
            type: object
        "500":
          description: 'error: string'
          schema:
            type: object
      summary: Получить список кораблей
      tags:
      - ships
    post:
      consumes:
      - application/json
      description: Add a new ship to the system
      parameters:
      - description: Ship data
        in: body
        name: ship
        required: true
        schema:
          $ref: '#/definitions/ds.Ship'
      produces:
      - application/json
      responses:
        "201":
          description: 'data: ds.Ship'
          schema:
            type: object
        "400":
          description: 'error: string'
          schema:
            type: object
        "500":
          description: 'error: string'
          schema:
            type: object
      summary: Создать корабль
      tags:
      - ships
  /api/ships/{id}:
    delete:
      description: Remove a ship from the system by ID
      parameters:
      - description: Ship ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: string'
          schema:
            type: object
        "400":
          description: 'message: string'
          schema:
            type: object
        "500":
          description: 'error: string'
          schema:
            type: object
      summary: Удалить корабль
      tags:
      - ships
    get:
      description: Retrieve details of a specific ship by ID
      parameters:
      - description: Ship ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: ds.Ship'
          schema:
            type: object
        "400":
          description: 'error: string'
          schema:
            type: object
        "404":
          description: 'error: string'
          schema:
            type: object
      summary: Один корабль
      tags:
      - ships
    put:
      consumes:
      - application/json
      description: Update details of an existing ship by ID
      parameters:
      - description: Ship ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated ship data
        in: body
        name: ship
        required: true
        schema:
          $ref: '#/definitions/ds.Ship'
      produces:
      - application/json
      responses:
        "200":
          description: 'data: ds.Ship'
          schema:
            type: object
        "400":
          description: 'error: string'
          schema:
            type: object
        "500":
          description: 'error: string'
          schema:
            type: object
      summary: Обновить поля корабля
      tags:
      - ships
  /api/ships/{id}/add-to-ship-bucket:
    post:
      description: Add a ship to a user's request draft
      parameters:
      - description: Ship ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: string, data: {request_ship_id: int, ship_id: int}'
          schema:
            type: object
        "400":
          description: 'message: string'
          schema:
            type: object
        "404":
          description: 'message: string'
          schema:
            type: object
        "500":
          description: 'status: string, description: string'
          schema:
            type: object
      summary: Добавить корабль в заявку
      tags:
      - ships
  /api/ships/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image for a specific ship
      parameters:
      - description: Ship ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file
        in: formData
        name: file
        required: true
        type: file
      - description: Image file (alternative)
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 'data: {ship_id: int, photo_url: string, message: string}'
          schema:
            type: object
        "400":
          description: 'message: string'
          schema:
            type: object
        "404":
          description: 'message: string'
          schema:
            type: object
        "500":
          description: 'message: string'
          schema:
            type: object
      summary: Добавить изображение
      tags:
      - ships
...
securityDefinitions:
  BearerAuth:
    description: 'Введите: Bearer {ваш_JWT_токен}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"